transfer data 1차 완료

+++ livedronemap-converter/src/main/java/gaia3d/domain/ImageMosaic.java
@@ -21,6 +21,6 @@ public class ImageMosaic {
 	private String image_dt;
 	
 	// 프로젝트 ID 
-	private Long project_id;
+	private Integer project_id;
 	
 }
